refaktorering del 1
