+++ src/main/kotlin/no/nav/fo/veilarbvarsel/brukernotifikasjonclient/producers/BrukernotifikasjonDoneProducer.kt
@@ -1,12 +1,12 @@
-package no.nav.fo.veilarbvarsel.brukernotifikasjon.producers
+package no.nav.fo.veilarbvarsel.brukernotifikasjonclient.producers
 
 import no.nav.brukernotifikasjon.schemas.Done
 import no.nav.brukernotifikasjon.schemas.Nokkel
 import no.nav.brukernotifikasjon.schemas.builders.DoneBuilder
 import no.nav.fo.veilarbvarsel.config.KafkaEnvironment
-import no.nav.fo.veilarbvarsel.kafka.KafkaProducerWrapper
-import no.nav.fo.veilarbvarsel.kafka.utils.KafkaCallback
-import no.nav.fo.veilarbvarsel.kafka.utils.props.KafkaAvroProducerProperties
+import no.nav.fo.veilarbvarsel.config.kafka.KafkaProducerWrapper
+import no.nav.fo.veilarbvarsel.config.kafka.utils.KafkaCallback
+import no.nav.fo.veilarbvarsel.config.kafka.utils.props.KafkaAvroProducerProperties
 import java.time.LocalDateTime
 
 class BrukernotifikasjonDoneProducer(
