+++ src/main/kotlin/no/nav/fo/veilarbvarsel/config/ApplicationContext.kt
@@ -2,10 +2,10 @@ package no.nav.fo.veilarbvarsel.config
 
 import io.micrometer.prometheus.PrometheusConfig
 import io.micrometer.prometheus.PrometheusMeterRegistry
-import no.nav.fo.veilarbvarsel.brukernotifikasjon.BrukernotifikasjonService
-import no.nav.fo.veilarbvarsel.brukernotifikasjon.producers.BrukernotifikasjonBeskjedProducer
-import no.nav.fo.veilarbvarsel.brukernotifikasjon.producers.BrukernotifikasjonDoneProducer
-import no.nav.fo.veilarbvarsel.brukernotifikasjon.producers.BrukernotifikasjonOppgaveProducer
+import no.nav.fo.veilarbvarsel.brukernotifikasjonclient.BrukernotifikasjonClient
+import no.nav.fo.veilarbvarsel.brukernotifikasjonclient.producers.BrukernotifikasjonBeskjedProducer
+import no.nav.fo.veilarbvarsel.brukernotifikasjonclient.producers.BrukernotifikasjonDoneProducer
+import no.nav.fo.veilarbvarsel.brukernotifikasjonclient.producers.BrukernotifikasjonOppgaveProducer
 import no.nav.fo.veilarbvarsel.dabevents.DabEventConsumer
 import no.nav.fo.veilarbvarsel.dabevents.DabEventProducer
 import no.nav.fo.veilarbvarsel.dabevents.DabEventService
@@ -37,7 +37,7 @@ class ApplicationContext {
     )
 
     val dabEventService = DabEventService(dabEventProducer)
-    val brukernotifikasjonService = BrukernotifikasjonService(beskjedProducer, oppgaveProducer, doneProducer)
+    val brukernotifikasjonService = BrukernotifikasjonClient(beskjedProducer, oppgaveProducer, doneProducer)
     val varselService = VarselService(dabEventProducer, brukernotifikasjonService, metrics)
 
     val dabEventConsumer = DabEventConsumer(
