+++ src/main/kotlin/no/nav/fo/veilarbvarsel/config/kafka/KafkaProducerWrapper.kt
@@ -1,6 +1,6 @@
-package no.nav.fo.veilarbvarsel.kafka
+package no.nav.fo.veilarbvarsel.config.kafka
 
-import no.nav.fo.veilarbvarsel.kafka.utils.KafkaCallback
+import no.nav.fo.veilarbvarsel.config.kafka.utils.KafkaCallback
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.slf4j.LoggerFactory
