+++ src/main/kotlin/no/nav/fo/veilarbvarsel/config/system/features/BackgroundJob.kt
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbvarsel.system.features
+package no.nav.fo.veilarbvarsel.config.system.features
 
 import io.ktor.application.*
 import io.ktor.util.*
