+++ src/main/kotlin/no/nav/fo/veilarbvarsel/config/system/features/ClosableJob.kt
@@ -1,0 +1,5 @@
+package no.nav.fo.veilarbvarsel.config.system.features
+
+import java.io.Closeable
+
+interface ClosableJob : Closeable, Runnable
