+++ src/main/kotlin/no/nav/fo/veilarbvarsel/main.kt
@@ -4,42 +4,20 @@ import com.fasterxml.jackson.databind.SerializationFeature
 import io.ktor.application.*
 import io.ktor.features.*
 import io.ktor.jackson.*
-import io.ktor.metrics.micrometer.*
-import io.ktor.routing.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
-import io.micrometer.core.instrument.binder.jvm.ClassLoaderMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmGcMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmMemoryMetrics
-import io.micrometer.core.instrument.binder.jvm.JvmThreadMetrics
-import io.micrometer.core.instrument.binder.system.FileDescriptorMetrics
-import io.micrometer.core.instrument.binder.system.ProcessorMetrics
-import io.micrometer.core.instrument.binder.system.UptimeMetrics
 import no.nav.fo.veilarbvarsel.config.ApplicationContext
-import no.nav.fo.veilarbvarsel.system.features.BackgroundJob
-import no.nav.fo.veilarbvarsel.system.healthApi
-import no.nav.fo.veilarbvarsel.varsel.varselApi
+import no.nav.fo.veilarbvarsel.config.system.features.BackgroundJob
+import no.nav.fo.veilarbvarsel.config.system.healthModule
 
 fun main() {
-    val port = System.getenv("PORT")?.toInt() ?: 8080
+    val port = System.getenv("SERVER_PORT")?.toInt() ?: 8080
     val server = embeddedServer(Netty, port, module = Application::mainModule)
     server.start()
 }
 
 fun Application.mainModule(appContext: ApplicationContext = ApplicationContext()) {
-    install(MicrometerMetrics) {
-        registry = appContext.metrics
-
-        meterBinders = listOf(
-            ClassLoaderMetrics(),
-            JvmMemoryMetrics(),
-            JvmGcMetrics(),
-            ProcessorMetrics(),
-            JvmThreadMetrics(),
-            FileDescriptorMetrics(),
-            UptimeMetrics(),
-        )
-    }
+    healthModule(appContext)
 
     install(ContentNegotiation) {
         jackson {
@@ -49,14 +27,6 @@ fun Application.mainModule(appContext: ApplicationContext = ApplicationContext()
 
     install(BackgroundJob.BackgroundJobFeature("DAB Events Consumer")) {
         job = appContext.dabEventConsumer
-    }
-
-    routing {
-        trace {
-            application.log.debug(it.buildText())
-        }
-        healthApi(appContext.metrics)
-        varselApi(appContext.dabEventService)
     }
 
     configureShutdownHook(appContext)
