--- src/main/kotlin/no/nav/fo/veilarbvarsel/system/features/ClosableJob.kt
@@ -1,5 +1,0 @@
-package no.nav.fo.veilarbvarsel.system.features
-
-import java.io.Closeable
-
-interface ClosableJob : Closeable, Runnable
