+++ src/main/kotlin/no/nav/fo/veilarbvarsel/varsel/VarselService.kt
@@ -1,12 +1,10 @@
 package no.nav.fo.veilarbvarsel.varsel
 
 import io.micrometer.prometheus.PrometheusMeterRegistry
-import no.nav.fo.veilarbvarsel.brukernotifikasjon.BrukernotifikasjonService
+import no.nav.fo.veilarbvarsel.brukernotifikasjonclient.BrukernotifikasjonClient
 import no.nav.fo.veilarbvarsel.dabevents.CreateVarselPayload
 import no.nav.fo.veilarbvarsel.dabevents.DabEventProducer
-import no.nav.fo.veilarbvarsel.kafka.utils.KafkaCallback
-import no.nav.fo.veilarbvarsel.varsel.domain.Varsel
-import no.nav.fo.veilarbvarsel.varsel.domain.VarselType
+import no.nav.fo.veilarbvarsel.config.kafka.utils.KafkaCallback
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.net.URL
@@ -14,7 +12,7 @@ import java.util.*
 
 class VarselService(
     private val dabEventProducer: DabEventProducer,
-    private val brukernotifikasjon: BrukernotifikasjonService,
+    private val brukernotifikasjon: BrukernotifikasjonClient,
     private val metrics: PrometheusMeterRegistry
 ) {
