--- src/main/kotlin/no/nav/fo/veilarbvarsel/varsel/domain/Done.kt
@@ -1,8 +1,0 @@
-package no.nav.fo.veilarbvarsel.varsel.domain
-
-data class Done(
-    val id: String,
-    val system: String,
-    val fodselsnummer: String,
-    val groupId: String,
-)
